更新ORM架构结构
