+++ lib/application.dart
@@ -2,15 +2,15 @@
  * @Description: 
  * @Author: chenzedeng
  * @Date: 2021-05-21 22:59:39
- * @LastEditTime: 2021-07-02 16:56:04
+ * @LastEditTime: 2021-07-11 22:23:04
  */
 
 import 'package:event_bus/event_bus.dart';
 import 'package:fluro/fluro.dart';
 import 'package:flutter/material.dart';
-import 'package:xy_music_mobile/service/player_service.dart';
 import 'package:xy_music_mobile/service/service_register.dart' as msr;
 import 'package:xy_music_mobile/config/theme_data.dart' as tdr;
+import 'package:xy_music_mobile/config/database_config.dart' as database;
 
 class Application {
   //Router全局路由管理对象
@@ -20,8 +20,6 @@ class Application {
   static final EventBus eventBus = EventBus();
 
   static late BuildContext context;
-
-  // static late final PlayerService playerService = PlayerService();
 
   ///IOS 可侧滑返回的跳转界面
   static Future navigateToIos(BuildContext context, String path,
@@ -55,11 +53,11 @@ class Application {
   static void applicationInit() {
     //音乐服务注册
     msr.register();
+
     //主题服务数据注入
     tdr.register();
-    //初始化音乐播放服务
-    // if (playerService == null) {
-    //   playerService = PlayerService();
-    // }
+
+    ///初始化DataBase
+    database.initDataBase();
   }
 }
